// src/funkoApp/enums/funkoFranchise.ts
export enum FunkoFranchise {
  Marvel = 'Marvel',
  DCComics = 'DC Comics',
  Disney = 'Disney',
  StarWars = 'Star Wars',
  HarryPotter = 'Harry Potter',
  Pokemon = 'Pokemon',
  GameOfThrones = 'Game of Thrones',
  RickAndMorty = 'Rick and Morty',
  StrangerThings = 'Stranger Things',
  TheWalkingDead = 'The Walking Dead',
  TheOffice = 'The Office',
  Friends = 'Friends',
  TheSimpsons = 'The Simpsons',
  SouthPark = 'South Park'
}